Added new double and int extensions for locations (miles-to-meters, distance between lat/long values).
Updated the AppBarToggleButton grouping helper to support when a button has been unchecked to make sure that if no button is checked, the last known state persists.

Changed the DeviceTrigger to include a property to support continuum for phone or not.

Updated the template for the ValidatingTextBox control to display the correct clear button symbol and tooltip.

Updated the sample app to separate components into separate views to easily extend and show how the components work)